# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Amar Steps - Swagger DOC ",
        "description": "L6 Swagger OpenApi description",
        "contact": {
            "email": "nshakib.se@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "127.0.0.1:8000",
            "description": "Sandbox OpenApi dynamic host server"
        }
    ],
    "paths": {
...
        "/api/v1/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get auth information",
                "description": "Returns auth data",
                "operationId": "profile",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieve auth profile.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/{id}": {
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Delete user.",
                "description": "User has been deleted successfully.",
                "operationId": "destroy-user",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User has been deleted successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "success": true,
                                        "data": "",
                                        "message": "User has been deleted successfully."
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Failed to delete.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "success": false,
                                        "message": "Failed to delete."
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}